New HRI framework/new_framework/requirements.txt
networkx==2.2
#dlib==19.13.0
six==1.12.0
pynput==1.4.2
scipy==1.2.2
Keras==2.3.1
numpy==1.16.5
llvmlite==0.31.0
librosa==0.7.1
matplotlib==2.2.4
paho-mqtt==1.4.0
requests-futures==1.0.0
#tensorflow==1.14.0


# only if using speech to text
google-cloud-speech==1.1.0

# it is good to install OpenCV manually but pip should also work
# opencv-python==3.3.0.9

#for pepper only install qi manually pip can not work
#qi
